代码生成其他模块

+++ mall-mbg/src/main/java/com/zscat/mallplus/marking/entity/SmsGroupMember.java
@@ -1,11 +1,12 @@
 package com.zscat.mallplus.marking.entity;
 
-import com.baomidou.mybatisplus.annotation.TableName;
 import com.baomidou.mybatisplus.annotation.IdType;
-import com.baomidou.mybatisplus.annotation.TableId;
-import java.util.Date;
 import com.baomidou.mybatisplus.annotation.TableField;
+import com.baomidou.mybatisplus.annotation.TableId;
+import com.baomidou.mybatisplus.annotation.TableName;
+
 import java.io.Serializable;
+import java.util.Date;
 
 /**
  * <p>
@@ -46,7 +47,7 @@ public class SmsGroupMember implements Serializable {
     private Integer status;
 
     @TableField("order_id")
-    private Integer orderId;
+    private Long orderId;
 
 
     public Long getId() {
@@ -113,11 +114,11 @@ public class SmsGroupMember implements Serializable {
         this.status = status;
     }
 
-    public Integer getOrderId() {
+    public Long getOrderId() {
         return orderId;
     }
 
-    public void setOrderId(Integer orderId) {
+    public void setOrderId(Long orderId) {
         this.orderId = orderId;
     }
